Separate Query implementation from interface
